# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sager Drone Tracking API",
        "description": "API for tracking and managing drones via MQTT",
        "version": "1.0.0"
    },
    "paths": {
        "/api/auth/register": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Register a new user",
                "operationId": "d764dd091cc4494ae0baf360b03319f3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "John Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "john@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password123"
                                    },
                                    "role": {
                                        "type": "string",
                                        "enum": [
                                            "user",
                                            "admin"
                                        ],
                                        "example": "user"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "User registered successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/auth/login": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Login user",
                "operationId": "8dcb70df1020986038d098cc08d05dae",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "password"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "admin@sager.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "admin123"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Login successful"
                    },
                    "401": {
                        "description": "Invalid credentials"
                    }
                }
            }
        },
        "/api/auth/logout": {
            "post": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Logout user",
                "operationId": "69281b12abb272c76871f19cb17ca563",
                "responses": {
                    "200": {
                        "description": "Logout successful"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/auth/me": {
            "get": {
                "tags": [
                    "Authentication"
                ],
                "summary": "Get authenticated user",
                "operationId": "b9abb1a7a74670a19c215c2c133f14d8",
                "responses": {
                    "200": {
                        "description": "User details"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/drones": {
            "get": {
                "tags": [
                    "Drones"
                ],
                "summary": "Get all drones",
                "operationId": "ca6c7156433adac5befeb122c5c40648",
                "parameters": [
                    {
                        "name": "serial",
                        "in": "query",
                        "description": "Filter by serial number (partial match)",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Items per page",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "default": 15
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of drones"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/drones/online": {
            "get": {
                "tags": [
                    "Drones"
                ],
                "summary": "Get all online drones",
                "operationId": "4f8f93d439595464d4b4d7b9deb25044",
                "responses": {
                    "200": {
                        "description": "List of online drones with current location"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/drones/nearby": {
            "get": {
                "tags": [
                    "Drones"
                ],
                "summary": "Get drones within specified radius",
                "operationId": "30cade170c8c64b6137244b1827e519b",
                "parameters": [
                    {
                        "name": "latitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 31.9783
                        }
                    },
                    {
                        "name": "longitude",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "example": 35.8309
                        }
                    },
                    {
                        "name": "radius",
                        "in": "query",
                        "description": "Radius in kilometers",
                        "required": false,
                        "schema": {
                            "type": "number",
                            "format": "float",
                            "default": 5
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List of nearby drones"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/drones/{serial}/flight-path": {
            "get": {
                "tags": [
                    "Drones"
                ],
                "summary": "Get drone flight path as GeoJSON",
                "operationId": "d72c632a20529e71575b0d9e096e80f7",
                "parameters": [
                    {
                        "name": "serial",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Flight path in GeoJSON format"
                    },
                    "404": {
                        "description": "Drone not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/drones/dangerous": {
            "get": {
                "tags": [
                    "Drones"
                ],
                "summary": "Get all dangerous drones with reasons",
                "operationId": "caf00330abd5b15e547598bbd22f7fd9",
                "responses": {
                    "200": {
                        "description": "List of dangerous drones with classification reasons"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/drones/{serial}/mark-safe": {
            "post": {
                "tags": [
                    "Drones"
                ],
                "summary": "Mark a drone as safe (Admin only)",
                "operationId": "d0c4b3730daad81181b08ea8e4083045",
                "parameters": [
                    {
                        "name": "serial",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Drone marked as safe"
                    },
                    "403": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Drone not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "tags": [
        {
            "name": "Authentication",
            "description": "API endpoints for authentication"
        },
        {
            "name": "Drones",
            "description": "API endpoints for drone management"
        }
    ]
}